+++ tests/test_psycopg_dbapi20.py
@@ -7,7 +7,6 @@ from psycopg.conninfo import conninfo_to_dict
 
 from . import dbapi20
 from . import dbapi20_tpc
-from .test_connection import drop_default_args_from_conninfo
 
 
 @pytest.fixture(scope="class")
@@ -145,7 +144,7 @@ def test_connect_args(monkeypatch, pgconn, args, kwargs, want, setpgenv):
     setpgenv({})
     monkeypatch.setattr(psycopg.connection, "connect", fake_connect)
     conn = psycopg.connect(*args, **kwargs)
-    assert drop_default_args_from_conninfo(got_conninfo) == conninfo_to_dict(want)
+    assert conninfo_to_dict(got_conninfo) == conninfo_to_dict(want)
     conn.close()